Assignment8 - Containerize It/requirements.txt
colorama==0.4.4
Flask==2.0.1
Flask-SQLAlchemy==2.3.2
gevent==21.8.0
google-api-core==2.1.1
google-api-python-client==2.29.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
google-cloud==0.34.0
google-cloud-core==2.1.0
google-cloud-datastore==2.1.6
google-cloud-language==2.3.1
google-cloud-speech==2.10.0
google-cloud-storage==1.42.3
google-cloud-translate==3.6.0
google-cloud-vision==2.6.2
google-crc32c==1.3.0
google-resumable-media==2.1.0
google-trans-new==1.1.9
googleapis-common-protos==1.53.0
googletrans==3.1.0a0

httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
idna==2.10
importlib-metadata==4.11.3
isort==5.9.3
Jinja2==3.0.1
Js2Py==0.71
json2html==1.3.0
lazy-object-proxy==1.6.0
libcst==0.3.21
libretranslatepy==2.1.1
load==2020.12.3
lxml==4.6.3
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.2
oauthlib==3.2.0
packaging==21.0
Pillow==8.4.0
pipwin==0.5.1
platformdirs==2.3.0
plyer==2.0.0
proto-plus==1.19.5
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyjsparser==2.7.1
PyJWT==2.3.0
pyparsing==2.4.7
PyPrind==2.11.3
PySimpleGUI==4.53.0
pySmartDL==1.3.4
python-dateutil==2.8.2
python-tsp==0.2.1
requests==2.27.1
requests-oauthlib==1.3.1